+++ src/main/java/me/kruase/mipt/animals/Camel.java
@@ -1,0 +1,8 @@
+package me.kruase.mipt.animals;
+
+import me.kruase.mipt.animals.interfaces.food.Herbivore;
+import me.kruase.mipt.animals.interfaces.movement.Land;
+
+
+public class Camel implements Land, Herbivore {
+}
