+++ src/main/java/me/kruase/mipt/animals/Dolphin.java
@@ -1,0 +1,8 @@
+package me.kruase.mipt.animals;
+
+import me.kruase.mipt.animals.interfaces.food.FishEater;
+import me.kruase.mipt.animals.interfaces.movement.Waterfowl;
+
+
+public class Dolphin implements Waterfowl, FishEater {
+}
