+++ src/main/java/me/kruase/mipt/animals/Eagle.java
@@ -1,0 +1,8 @@
+package me.kruase.mipt.animals;
+
+import me.kruase.mipt.animals.interfaces.food.Predator;
+import me.kruase.mipt.animals.interfaces.movement.Flying;
+
+
+public class Eagle implements Flying, Predator {
+}
