+++ src/main/java/me/kruase/mipt/animals/Tiger.java
@@ -1,0 +1,8 @@
+package me.kruase.mipt.animals;
+
+import me.kruase.mipt.animals.interfaces.food.BeefEater;
+import me.kruase.mipt.animals.interfaces.movement.Land;
+
+
+public class Tiger implements Land, BeefEater {
+}
