+++ src/main/java/me/kruase/mipt/animals/interfaces/food/BeefEater.java
@@ -1,0 +1,7 @@
+package me.kruase.mipt.animals.interfaces.food;
+
+import me.kruase.mipt.food.Beef;
+
+
+public interface BeefEater extends Eater<Beef> {
+}
