+++ src/main/java/me/kruase/mipt/animals/interfaces/food/FishEater.java
@@ -1,0 +1,7 @@
+package me.kruase.mipt.animals.interfaces.food;
+
+import me.kruase.mipt.food.Fish;
+
+
+public interface FishEater extends Eater<Fish> {
+}
