+++ src/main/java/me/kruase/mipt/animals/interfaces/food/Herbivore.java
@@ -1,0 +1,7 @@
+package me.kruase.mipt.animals.interfaces.food;
+
+import me.kruase.mipt.food.Grass;
+
+
+public interface Herbivore extends Eater<Grass> {
+}
