+++ src/main/java/me/kruase/mipt/animals/interfaces/food/Predator.java
@@ -1,0 +1,7 @@
+package me.kruase.mipt.animals.interfaces.food;
+
+import me.kruase.mipt.food.abc.Meat;
+
+
+public interface Predator extends Eater<Meat> {
+}
