+++ src/main/java/me/kruase/mipt/animals/interfaces/movement/Flying.java
@@ -1,4 +1,4 @@
-package me.kruase.mipt_mts_homework.animals.interfaces.movement;
+package me.kruase.mipt.animals.interfaces.movement;
 
 public interface Flying {
     default void fly(int distance) {
