+++ src/main/java/me/kruase/mipt/animals/interfaces/movement/Waterfowl.java
@@ -1,4 +1,4 @@
-package me.kruase.mipt_mts_homework.animals.interfaces.movement;
+package me.kruase.mipt.animals.interfaces.movement;
 
 public interface Waterfowl {
     default void swim(int distance) {
