+++ src/main/java/me/kruase/mipt/food/abc/Food.java
@@ -1,4 +1,4 @@
-package me.kruase.mipt_mts_homework.food.abc;
+package me.kruase.mipt.food.abc;
 
 
 public abstract class Food {
