+++ src/main/java/me/kruase/mipt/food/abc/Meat.java
@@ -1,4 +1,5 @@
-package me.kruase.mipt_mts_homework.food.abc;
+package me.kruase.mipt.food.abc;
+
 
 public abstract class Meat extends Food {
     public Meat(int mass) {
