--- src/main/java/me/kruase/mipt_mts_homework/animals/Dolphin.java
@@ -1,8 +1,0 @@
-package me.kruase.mipt_mts_homework.animals;
-
-import me.kruase.mipt_mts_homework.animals.interfaces.food.FishEater;
-import me.kruase.mipt_mts_homework.animals.interfaces.movement.Waterfowl;
-
-
-public class Dolphin implements Waterfowl, FishEater {
-}
