--- src/main/java/me/kruase/mipt_mts_homework/animals/Eagle.java
@@ -1,8 +1,0 @@
-package me.kruase.mipt_mts_homework.animals;
-
-import me.kruase.mipt_mts_homework.animals.interfaces.food.Predator;
-import me.kruase.mipt_mts_homework.animals.interfaces.movement.Flying;
-
-
-public class Eagle implements Flying, Predator {
-}
