--- src/main/java/me/kruase/mipt_mts_homework/animals/Horse.java
@@ -1,8 +1,0 @@
-package me.kruase.mipt_mts_homework.animals;
-
-import me.kruase.mipt_mts_homework.animals.interfaces.food.Herbivore;
-import me.kruase.mipt_mts_homework.animals.interfaces.movement.Land;
-
-
-public class Horse implements Land, Herbivore {
-}
