--- src/main/java/me/kruase/mipt_mts_homework/animals/Tiger.java
@@ -1,8 +1,0 @@
-package me.kruase.mipt_mts_homework.animals;
-
-import me.kruase.mipt_mts_homework.animals.interfaces.food.BeefEater;
-import me.kruase.mipt_mts_homework.animals.interfaces.movement.Land;
-
-
-public class Tiger implements Land, BeefEater {
-}
