--- src/main/java/me/kruase/mipt_mts_homework/animals/interfaces/food/FishEater.java
@@ -1,7 +1,0 @@
-package me.kruase.mipt_mts_homework.animals.interfaces.food;
-
-import me.kruase.mipt_mts_homework.food.Fish;
-
-
-public interface FishEater extends Eater<Fish> {
-}
